+++ accounts/templates/login.html
@@ -3,50 +3,13 @@
     Login
 {% endblock title %}
 {% block body %}
-    <div class="flex min-h-full flex-col justify-center px-6 py-12 lg:px-8">
-        <div class="sm:mx-auto sm:w-full sm:max-w-sm">
-            <img class="mx-auto h-10 w-auto"
-                 src="https://tailwindcss.com/plus-assets/img/logos/mark.svg?color=indigo&shade=600"
-                 alt="Your Company" />
-            <h2 class="mt-10 text-center text-2xl/9 font-bold tracking-tight text-gray-900">Sign in to your account</h2>
-        </div>
-        <div class="mt-10 sm:mx-auto sm:w-full sm:max-w-sm">
-            <form class="space-y-6" action="{% url 'login' %}" method="POST">
-                {% csrf_token %}
-                <div>
-                    <label for="email" class="block text-sm/6 font-medium text-gray-900">Insira o seu Email</label>
-                    <div class="mt-2">
-                        <input type="email"
-                               name="username"
-                               id="email"
-                               autocomplete="email"
-                               required
-                               class="block w-full rounded-md bg-white px-3 py-1.5 text-base text-gray-900 outline-1 -outline-offset-1 outline-gray-300 placeholder:text-gray-400 focus:outline-2 focus:-outline-offset-2 focus:outline-indigo-600 sm:text-sm/6" />
-                    </div>
-                </div>
-                <div>
-                    <div class="flex items-center justify-between">
-                        <label for="password" class="block text-sm/6 font-medium text-gray-900">Senha</label>
-                        <div class="text-sm">
-                            <a href="#" class="font-semibold text-indigo-600 hover:text-indigo-500">Forgot password?</a>
-                        </div>
-                    </div>
-                    <div class="mt-2">
-                        <input type="password"
-                               name="password"
-                               id="password"
-                               autocomplete="current-password"
-                               required
-                               class="block w-full rounded-md bg-white px-3 py-1.5 text-base text-gray-900 outline-1 -outline-offset-1 outline-gray-300 placeholder:text-gray-400 focus:outline-2 focus:-outline-offset-2 focus:outline-indigo-600 sm:text-sm/6" />
-                    </div>
-                </div>
-                <div>
-                    <button type="submit"
-                            class="flex w-full justify-center rounded-md bg-indigo-600 px-3 py-1.5 text-sm/6 font-semibold text-white shadow-xs hover:bg-indigo-500 focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600">
-                        Sign in
-                    </button>
-                </div>
-            </form>
-        </div>
-    </div>
+    <h2>Login</h2>
+    <form action="{% url 'login' %}" method="POST">
+        {% csrf_token %}
+        <label for="email">Insira o seu Email</label>
+        <input type="email" name="email" required />
+        <label for="password">Senha</label>
+        <input type="password" name="password" required />
+        <button type="submit">Logar</button>
+    </form>
 {% endblock body %}
