+++ accounts/urls.py
@@ -3,11 +3,11 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path("login/", views.AccountLogin.as_view(), name="login"),
-    path("register/", views.AccountRegister.as_view(), name="register"),
-    path("logout/", views.AccountLogout.as_view(), name="logout"),
-    path("", views.AccountList.as_view(), name="account_list"),
-    path("<int:pk>", views.AccountDetail.as_view(), name="account_detail"),
-    path("<int:pk>/update", views.AccountUpdate.as_view(), name="account_update"),
-    path("<int:pk>/delete", views.AccountDelete.as_view(), name="account_delete"),
+    path("login/", views.login, name="login"),
+    path("register/", views.register, name="register"),
+    path("logout/", views.logout, name="logout"),
+    path("", views.account_list, name="account_list"),
+    path("<int:id>", views.account_detail, name="account_detail"),
+    path("<int:id>/update", views.account_update, name="account_update"),
+    path("<int:id>/delete", views.account_delete, name="account_delete"),
 ]
